+++ Source/EventFlow.EventStores.EventStore/EventStoreEventPersistence.cs
@@ -117,7 +117,7 @@ namespace EventFlow.EventStores.EventStore
                     })
                 .ToList();
 
-            var expectedVersion = Math.Max(serializedEvents.Min(e => e.AggregateSequenceNumber) - 1, 0);
+            var expectedVersion = Math.Max(serializedEvents.Min(e => e.AggregateSequenceNumber) - 2, ExpectedVersion.NoStream);
             var eventDatas = serializedEvents
                 .Select(e =>
                     {
@@ -131,9 +131,7 @@ namespace EventFlow.EventStores.EventStore
 
             try
             {
-                using (var transaction = await _connection.StartTransactionAsync(
-                    id.Value,
-                    expectedVersion == 0 ? ExpectedVersion.NoStream : expectedVersion)
+                using (var transaction = await _connection.StartTransactionAsync(id.Value, expectedVersion)
                     .ConfigureAwait(false))
                 {
                     await transaction.WriteAsync(eventDatas).ConfigureAwait(false);